Fixed tests broken after refactoring
